pyproject.toml
[project]
name = "gig-interveners"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"




dependencies = [
    "bpy>=4.0.0",
    "bvhio>=1.4.6",
    "ipython>=8.31.0",
    "ipywidgets>=8.1.5",
    "jupyterlab>=4.3.4",
    "librosa>=0.10.2.post1",
    "marimo>=0.10.12",
    "numpy>=1.26.4",
    "praat-parselmouth>=0.4.5",
    "pysoundfile>=0.9.0.post1",
    "tyro>=0.9.8",
    "moderngl-window==2.4.6",
    "ruff>=0.9.1",
    "wasabi>=1.1.3",
    "scipy>=1.15.1",
    "scikit-learn>=1.6.1",
    "rich>=13.9.4",
    "matplotlib>=3.10.0",
    "einops>=0.8.0",
    "pdbpp>=0.10.3",
    "rich-tools>=0.5.1",
    "emd>=0.8.0",
    "ipykernel>=6.29.5",
    "pyprojroot>=0.3.0",
    "omegaconf>=2.3.0",
    "zarr>=2.13",
    "pyts>=0.13.0",
    "joblib>=1.4.2",
    "seaborn>=0.13.2",
    "pyunicorn>=0.8.2",
    "plum-dispatch>=2.5.7",
    "numba>=0.61.0",
    "cython>=3.0.12",
    "jax[cuda12]>=0.5.3; sys_platform == 'linux'",
    "jax>=0.5.3; sys_platform == 'darwin'",
    "ott-jax>=0.5.0",
    "tslearn>=0.6.3",
    "polars>=1.26.0",
]


[tool.ruff]
line-length = 120

[tool.ruff.format]
indent-style = "space"
docstring-code-format = true


[tool.setuptools]
py-modules = []
